Initial commit of project structure
This is the initial commit of the project structure. It contains all the required .env files to get the porject started. It is recommended to remove these for actual projects. It is a docker container project with two sides, the front-end labelled ui, and the back-end labelled api. The api relies on a postgres database. Both ui and api are served via an ngix container.